+++ FrontEnd/src/pages/Settings/SettingsContext.js
@@ -22,6 +22,7 @@ export const SettingsProvider = ({ children }) => {
     const initialSettings = {
         debug: false,
         darkMode: true,
+        defaultModel: "gpt-4o-mini",
         aiColour: false,
         language: 'en',
         augmentedPromptsEnabled: "off",
@@ -108,6 +109,7 @@ export const SettingsProvider = ({ children }) => {
                     ...prevSettings,
                     debug: loadedConfig.interface.debug ?? initialSettings.debug,
                     darkMode: loadedConfig.interface.dark_mode ?? initialSettings.darkMode,
+                    defaultModel: loadedConfig.interface.default_model ?? initialSettings.defaultModel,
                     aiColour: loadedConfig.interface.ai_colour ?? initialSettings.aiColour,
                     augmentedPromptsEnabled:
                         loadedConfig.beta_features.augmented_prompts_enabled ?? initialSettings.augmentedPromptsEnabled,