+++ CHANGELOG.md
@@ -13,10 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generated files are displayed under the underneath the response
 - Costs for individual functionalities are displayed
 - ~~Option~~ to run write pages workflow in parallel
+- Select default model
 
 ### Changed
 
-- Summaries properly under files (disabled by default)
+- Summaries properly displayed under files (disabled by default)
 
 ### Removed